Reformat cancellation token package to separate source files and build outputs
