--- packages/cancellation-token/index.ts
@@ -1,2 +1,0 @@
-export * from "./lib/Token";
-export * from "./lib/TokenSource";
