+++ packages/cancellation-token/src/Token.ts
