+++ packages/cancellation-token/src/TokenSource.ts
